+++ app/src/main/java/nextstep/payments/ui/component/PaymentCard.kt
@@ -1,4 +1,4 @@
-package nextstep.payments.ui
+package nextstep.payments.ui.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
